+++ src/lib.rs
@@ -109,13 +109,13 @@ mod accounts;
 /// The Client establishes the connection to TWS or the Gateway.
 /// It manages the routing of messages between TWS and the application.
 pub mod client;
-mod constants;
+
 /// A [Contract](crate::contracts::Contract) object represents trading instruments such as a stocks, futures or options.
 ///
 /// Every time a new request that requires a contract (i.e. market data, order placing, etc.) is sent to the API, the system will try to match the provided contract object with a single candidate. If there is more than one contract matching the same description, the API will return an error notifying you there is an ambiguity. In these cases the API needs further information to narrow down the list of contracts matching the provided description to a single element.
 pub mod contracts;
-/// Describes primary data structures used by the model.
-pub(crate) mod domain;
+// Describes primary data structures used by the model.
+//pub(crate) mod domain;
 pub mod errors;
 /// APIs for retrieving market data
 pub mod market_data;
